requirements.txt
# Runtime dependencies only.
# Dev/test tools belong in requirements-dev.txt and requirements-test.txt.

# AWS integration
boto3>=1.34.0,<2.0.0
botocore>=1.34.0,<2.0.0

# Azure integration
azure-functions>=1.18.0,<2.0.0
azure-identity>=1.16.0,<2.0.0
azure-keyvault-secrets>=4.8.0,<5.0.0
azure-monitor-ingestion>=1.0.0,<2.0.0
azure-core>=1.30.0,<2.0.0

# Security & encryption
cryptography>=42.0.0,<46.0.0
PyJWT>=2.8.0,<3.0.0

# Configuration
PyYAML>=6.0.0,<7.0.0
watchdog>=4.0.0,<7.0.0

# Date/time
python-dateutil>=2.8.2,<3.0.0

# HTTP
requests>=2.31.0,<3.0.0
aiohttp>=3.9.0,<4.0.0

# Validation & observability
jsonschema>=4.20.0,<5.0.0
prometheus_client>=0.20.0,<1.0.0
